+++ packages/json-rpc/polkadot-sdk-compat/src/index.ts
@@ -1,0 +1,17 @@
+import { parsed } from "./parsed"
+import {
+  translate,
+  unpinHash,
+  patchChainHeadEvents,
+  fixUnorderedEvents,
+} from "./parsed-enhancers"
+
+export default parsed(
+  translate,
+  fixUnorderedEvents,
+  unpinHash,
+  patchChainHeadEvents,
+)
+
+export * from "./parsed"
+export { translate, fixUnorderedEvents, unpinHash, patchChainHeadEvents }